refactor: Create src/robots.ts
