+++ __tests__/index.ts
@@ -26,7 +26,6 @@ import {
   returnsText,
   givenUuid,
   currentTestEnvironment,
-  currentTestEnvironmentWhen,
   expectToBeRedirectTo,
   localTestEnvironment,
 } from './__utils__'
@@ -64,28 +63,4 @@ describe('Enforce HTTPS', () => {
 
     expect(await response.text()).toBe('content')
   })
-})
-
-test('Disallow robots in non-productive environments', async () => {
-  const env = currentTestEnvironmentWhen(
-    (config) => config.ENVIRONMENT !== 'production'
-  )
-
-  const response = await env.fetch({ subdomain: 'de', pathname: '/robots.txt' })
-
-  expect(await response.text()).toBe('User-agent: *\nDisallow: /\n')
-})
-
-test('Return explicit robots rules in production', async () => {
-  global.ENVIRONMENT = 'production'
-  const env = currentTestEnvironmentWhen(
-    (config) => config.ENVIRONMENT === 'production'
-  )
-
-  const request = new Request('https://de.serlo.org/robots.txt')
-  const response = await env.fetchRequest(request)
-  const text = await response.text()
-
-  expect(text).toContain('User-agent: *')
-  expect(text).toContain('Disallow: /backend')
 })
